pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Dan Sclearov <dsclearov@starnet.md>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.15"
django = "^4.1.3"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
